Fixed Motorola S-record reading - previously broken by a code "cleanup"
I had obviously tested the code with -d debug-srec, which works. But if
you try to run it normally, because sskip is defined *after* sheader, it
fails.

I had moved sskip to put the s<whatever> words into the same order in
both the debug and normal code. But failed to test that the normal code
no longer compiled.