Fixed HC08 USB chat code - updated to "chunk-style"
I had changed the interact to read and write in chunks - always. I
updated the chat2 and chat3 serial code, but had neglected to fix the
chat USB code.

Now fixed!

But it becomes clear that the code that lives on the target to flash a
chunk should clear the FSTAT byte and return the value of FSTAT at the
end... Maybe I should fix that.