msp430: Fixed remote execution bug!
In my zeal to convert from using cfa's to pfa's (as the result of
' <target-word> ) I inadvertently broke remote execution. I forgot *one*
place where a cfa needed to be converted into a pfa: right before
passing it to the target to execute! So, predictably, the target went
into the weeds.

Now fixed!