msp430: Disassembler smarter about code fields! Now prints names!
If the disasm finds something that looks like a code word's code field -
ie a cell that points to the following cell - *and* it finds a matching
entry in the .target-runtime. chain (which links to the .target.
chain) it will print

code <name>

instead of interpreting the code field as an instruction and, more often
than not, getting confused.

To make these code annotations stand out, regular instruction printing
has been indented two spaces.