msp430: ' returns cfa (again)
This is a partial reversion of commit
5b3e8e423e48aa788de452c0b5e154e46d0557de.

As an experiment, and because it made sense for RISC-V, that commit also
changed the MSP430 target to store pfa's rather than cfa's as the
address for each target word.

One reason I did this was to make disassembly less painful, but because
disassembly now has a fancy feature of annotating code words with their
names, this is now unnecessary.

I also checked in with the community, by looking at standards -
including the Forth 2012 draft standard. It says there that ' should
return an xt (which for the MSP430 is a cfa), such that

  ' <word> execute

has the identical behaviour to

  <word>

This change reverts this behaviour. The target addresses stored in the
host dictionary are cfa's again, and ' returns the cfa.