core: Catch/throw working again!
The correct design of this was tricky. (And it's not entirely clear to
me that this is indeed the *correct* design. ;-)

Commit a75257f40309 broke catch and throw. The trouble? Before that
commit, abort would print print the location of the error (file name and
line number) and *then* unwind the stack.

Because I wanted to move toward being able to use catch and throw not
simply to reset the system, but also to catch errors (eg timeouts) and
respond to them (by retrying, by trying a different device, etc), I
postponed the error reporting until *after* the stack was unwound.

The trouble with *that*? If we are loading a file, and that file loads
another file, several things happen. First, we save the old file name
(for oddball reasons it's stored in the variable "zloading") and the
current line number (which lives in "@line"). These old values are
pushed onto the R stack as special frames, zloading is set to the name
of the file to load, and @line is set to 0.

When we are finished reading everything, these stack frames get popped
and the variables are reset to their previous values. When an exception
(eg, throw) occurs, a similar stack "unwinding" occurs, and the
variables are reset in the same way.

When we are sitting at the command line, zloading is empty, and @line is
0. So if an error occurs during file loading and we unwind the stack and
*then* try to print the location of the error, it will always simply say
that the error was "interactive". This is obviously not what we want, so
we have to print the error before unwinding, like we did before.

The difference is that now the error printing is under the control of
the code that calls catch. What happens is this: When catch executes, it
pushes a special catch frame and links it onto a chain of catch frames.
It "fetches" the cfa of the following word - unforthy, yes, but it saves
us doing ['] foobar - pushes as its return address the address
*following* that cfa, and then executes the cfa.

If nothing bad happens, execute returns normally, the catch frame is
popped and unlinked, and we go on our merry way.

If an exception occurs and is thrown, throw does only thing: it jumps
directly to the return address that catch pushed, but doesn't touch the
R stack. All the nested frames are still there, as is the catch frame.
It is the responsibility of the code that follows the catch to unwind
the stack, print an error, or whatever it wants to do.

The Forth interactive loop code prints the error (with file name and
line number) and then unwinds the stack, passing a flag (which is
currently -1) to unwind. The cleanup words that execute to undo each
frame on the R stack can interpret this undwinding flag however they
wish. Currently it is used to "silence" noisy errors (such as stack
depth mismatches) that would distract from the readability of the syntax
error that is the real problem. But it could be used for whatever
purpose. The only complication of "overloading" this flag is that every
cleanup routine that might execute needs to understand and interpret the
flag bits correctly - so adding flag bits can be a bit of work.

This setup works well, even though it seems to me to be a bit
bass-ackwards. Hopefully I won't have to revisit the design yet again
though - I'm tired of thinking about it. ;-)