msp430: Added beginnings of support for MSP430FR6989
Right now I can successfully *compile* the chat code for the FR6989, but
it doesn't run. (I'm testing on the Launchpad that has that chip on it.)

I had to make a lot of changes to make this work. Even though the G2553
(our previous sole MSP430 target) and the FR6989 are "MSP430" devices,
they are different enough that several key parts of the chat code had to
be modified.

Rather than write little routines for each supported, I decided to try
doing it "declaratively". The resulting code is a bit ugly and hard to
read, but it should make it comparatively easy to add chips to the
supported list (at least as far as chat is concerned).

This new code generates the same binary image for the G2553, so it
shouldn't break existing chips.

Also, now you have to specify, any time you load MSP430 code, what kind
of chip to compile for. Right now you have two options:

  -d g2553   ( MSP430G2553)
  -d fr6989  ( MSP430FR6989)