msp430: Modified flash code to work with FRAM devices
It's now possible to "flash" the Forth kernel onto the FR6989 and
remotely execute Forth words on it! The incantation is as usual:

  ./muforth -d fr6989 -f target/MSP430/load-kernel.mu4
  chat
  flash-image
  verify

For now, verify is a little noisier: it prints the address ranges it is
verifying, followed by a colon; if no addresses *follow* the colon, then
everything matches.

While making these changes, I also dramatically simplified how the
normal flashing process works, esp when the chat code and vectors are
involved. Hopefully nothing broke. ;-)

Also, it became clear when executing Forth code that the UART connection
provided by the ez-FET interface on the board, even though it is
nominally running at 115,200 bps, is *very* slow. Executing a Forth word
remotely takes a second or more. It feels dramatically more sluggish
than doing exactly the same thing but using a different USB UART device
- eg, a PL2303.

I plan to make some changes to mitigate this, but I wish TI had done a
better job on their USB UART implementation!