msp430: It is now possible to re-flash the fr6989 chat from muforth
load-chat had a problem: it assumed that all MSP devices have ram
starting at 0200. This is the middle of the i/o space on the FR6989.

Also, the chat code had a mismatched .if/.then, which caused it to
compile some code that didn't belong in the "in-ram" version of chat.

For the brave:

As always, the commands to re-flash the chat code go something like
this:

  ./muforth -d fr6989 -d in-ram -f target/MSP430/load-chat.mu4
  chat
  reset

The PC should now show a value in the 1c00 range (RAM starts at 1c00 on
this device).

^C and reload the flash-resident version:

  ./muforth -d fr6989 -f target/MSP430/load-chat.mu4
  chat
  flash-image
  verify

Press the reset button on the board and "chat" again and you should see
the PC move back up to the fe00 range.