msp430: Speed up remote exec by reading/writing stack in big chunk
This makes some difference, but I think by changing the chat code I can
do even better.

I was thinking... Maybe the slow/bad performance of the debug USB UART
is not TI's fault! The UART is a USB class-compliant implementation that
enumerates as a /dev/ttyACMx, unlike the PL2303 and FTDI USB UARTs -
which have their own drivers and enumerate as /dev/ttyUSBx devices.

Maybe the "custom" drivers are tuned for reasonable performance, whereas
the Linux kernel USB-CDC ACM drivers (yes, that's where I was testing
this) are terrible?

I should test on the Mac to see if it's better there.

And trying this on a BSD machine might be interesting too...