msp430: A new version of the chat code?
Calling this v3. The idea is to make it simpler and faster. Faster
mostly because it passes big chunks back and forth between host and
target, and for some USB UARTs the turn-around takes a long time.

This version consists of an MSP430 assembler chat/debug stub, and
corresponding host-side code. By going through contortions it's possible
to connect the two. I've tested that the basic commands work. (I ran it
in RAM on the target.)

There is no Flash support. This is, at least for now, considered an
FRAM-only version of chat. Flash support could be added back in if it
seems like this way of doing things is a win.

There is also no "paranoid" support. No checksums. It just plays fast
and loose with the truth. This has not been a problem with other serial
chat implementations, so I don't expect it will be one here.

I plan to hook it into the interact code shortly. The overarching goal:
I'm trying to make remote exec on the FR6989 Launchpad fast enough that
you don't feel the lag. Right now it's kind of a disaster.