msp430: Tweaks to chat v3 code; works with Forth kernel too!
Remote execution is much faster than before. There is a tiny lag, but
it's nothing like it was with the original v2 chat.

Loading the v3 code is not currently the default. You can however easily
test it out. Try this:

First, reset the board. Then load the new chat code into RAM and start
it up:

  ./muforth -d fr6989 -d chat-v3 -d in-ram target/MSP430/load-chat.mu4
  chat
  go

Then load the Forth kernel code (which you should already have flashed):

  ./muforth -d fr6989 -d chat-v3-host target/MSP430/load-kernel.mu4
  chat

And then try something:

  1 2 +

I'm probably going to make chat-v3 the default for the fr6989 in a
future commit.