[hc08] Flash verify warns when flash empty, explains chat commit
Flash verification is now more verbose in two situations:

(1) If you have not compiled any code into the flash region, verify will
warn that the region is empty. This is usually a mistake - check your
command line arguments!

(2) If we are loading the chat code, the first four bytes of the image
might differ because that is where the commit that identifies the chat
version is stored. As long as there are no other differences, the code
is fine and does not need re-flashing.