[core] Console colors can be used on ChromeOS!
A while back I wrote some code for farting around with Linux text
console colors. (I have an old laptop that I use exclusively in text
mode, using a framebuffer console.)

I wanted to be able to make similar changes to the colors used by hterm,
the terminal emulator in ChromeOS.

I'm not suggesting that the generated colors are *worth* using, but if
you want to try this code, do this:

  ./muforth -f lib/console-palette.mu4
  gen
  export-json

Then select the two lines of text that get generated (an array of #rrggbb
values), open the options tab for hterm (the easiest way is to type
Ctrl-Shift-P while in a terminal window) and paste the copied array into
the box entitled "color-palette-overrides". You should also change the
foreground color to match the eighth color in the array.

(This is also how one might switch to the Solarized color scheme.)

Unless you also check the "enable-bold-as-bright" box (which is
mis-named), hterm will only use the first eight colors, and will bold
them to make them bright. By enabling "bold-as-bright" you actually
enable *both* the use of colors 9 to 16 *and* bold-as-bright. Which is a
bit lame.

Another thing that I don't understand yet about hterm is how I end up
with TERM set to "xterm", when, very clearly, the hterm options tab is
setting it to "xterm-256color" instead.

Oh, and you get blinking text (eg from weirdly set DIRCOLORS) even when
the "enable-blink" box is unchecked.

There will always be ChromeOS mysteries, I guess.