Refactored editline so that other code can tap into it
I wanted to create a "teleprompter" - a way to paste pre-configured
command lines, one at a time, into the command line buffer. I also
needed a way to add a keypress to do this.

There is now a defer'ed word - cl-key-unmatched - that is called when
the key pressed has an ASCII value of < 32 (it's a control char) but is
unused by the editline code. In this way client code can match unused
control chars and do something with them.

Also, the code to "paste" a command line into the buffer - now called
"cl-show" was refactored, both to make it easy for outside code, but
also to clean up redundancy in the editline code.