Teleprompter FTW!!1!1!
This makes it possible to "pre-record" a sequence of command lines,
which can be "commentary text" or instructions to a human reader, or a
series of muforth command lines.

I wrote it to make the creation of screencasts less painful for the
viewer - and to reduce the "performance anxiety" that goes along with
the "live coding" aspect of screencasting.

But I think this has interesting tutorial value as well, for someone
sitting at a muforth command line and wondering what to do next.

I've created a muforth subdirectory - prompt/ - where I intend to
collect tutorial and screencasting-related teleprompter sequences.