msp430: Fixed copying of ram to target; now copies >512 bytes!
When trying to load a fair amount of code into RAM, the target would
mysteriously stop working.

It turns out, because the "word count" value in the serial protocol is
only a byte, that copying >512 bytes across to the target fails in
confusing ways.

The simple solution: write the RAM contents to the target in chunks of
256 bytes. (This was copied from target/RISCV/interact.mu4. I had a very
similar problem there.)