msp430: Improved chat session reliability
If you run some code on the target that takes too long to execute, the
connection will time out, and you'll see the message "<key timed out".
From this point it was hard to resync the host to the target, even
though nothing was necessarily messed up on either side.

I've re-factored some of the "interact" code to make it possible to
reliably reconnect in this situation. Just type

  rechat

and the host should query the target's status, fetch and print the
registers, and be ready for more interaction.

I've tested with with both the chatv2 and chatv3 code.