[core] Updated src/lfsr.c, the (experimental) LFSR code
An LFSR is a linear feedback shift register. They are useful for
computing CRCs and quickly computing pseudo-random bit sequences. A
"good" LFSR that is n bits long has (2^n)-1 states - that is, it visits
every possibile configuration of 2^n bits except for the all-zero state
before repeating. The all-zero state is a "stuck state" for LFSRs - they
never leave it. (XORing zeros with other zeros just creates more zero.
You need at least one 1 to get things going - and one 1 is enough to
generate the entire sequence!)

The implementations in src/lfsr.c predated the conversion of muforth to
being a "64-bit everywhere" system, so I tweaked a few things to get it
to compile and run again.

One interesting/terrible thing that I learned about C is this: the
compiler treats literals as _signed int_ quantities (apparently) so that
if you do this:

  uint64_t x = (1 << 31)

x will not contain what I would expect (the value 0x0000_0000_8000_0000)
but instead will contain the outrageous value 0xffff_ffff_8000_000! C
has helpfully sign-extended a non-negative integer before assigning it
to an unsigned value! The way to prevent this frankly idiotic (and
doubtless error-producing) behaviour is to annotate the 1 as an unsigned
value, thus:

  uint64_t x = (1U << 31)

A similarly odd quirk is that you can't do this:

  uint64_t x = (1 << 63)

because the resulting value doesn't fit into an "int", which is the
default "container" that the compiler seems to want to put that 1 into.
Instead, we have to say that it's a long value. Just to be safe, we
might as well annotate it as unsigned too:

  uint64_t x = (1UL << 63)

I tripped over both of these with this code. Kinda crazy.

It's my understanding that this is one of annoyances of C that the
designers of Go tried to fix.

Oh and by the way... The 64-bit LFSR defined in src/lfsr.c seems to have
odd behaviour. Maybe it's correct, maybe not! (The taps that is uses are
standard ones.) But you really have to run it for a while (by generating
several tens of kilobits) before it stops generating huge swaths of
consecutive zeros...

I should also mention that this code is *not* compiled by default (which
is why I didn't know that muforth's upgrade to 64 bits had broken it).
To compile it you have create a file in src/ called local.mk containing
the following line:

  WITH_LFSR=1

and then run make.