msp430: Fixed (and improved) LFSR code
I got lucky before. I was loading ffff_ffff as the starting point, and
running the LFSR, counting the number of states that it passed thru
before it returned to the value ffff_ffff.

One problem: the previous code initialized the state like this:

  "ffff # y mov  "ffff x mov

The " prefix means hex; the # operator means immediate. Note that
loading x *lacks* the #. The code it compiled was actually an indirect
read using the z register... But, by luck, every time I ran this code, z
pointed to unprogrammed flash or non-existent memory containing fffff...

I'm now initializing x and y with 1 and 0, resp. This starts the LFSR at
0000_0001. And the loop exit condition is the same: x == 1 and y == 0.