[core and msp430]: Fun with CRC16-CCITT!!
Rabbit hole alert.

I thought it might be a good idea to be able to test changes to my
MSP430 chat and flashing code using the BSL. I haven't touched the BSL
code for a long time, and the FR6989 uses a completely different BSL
than the G2553. (Yeah, uh.. Thanks, TI!!)

Among other things, it uses CRC16-CCITT for checksums.

So I went online searching for info about CRCs - a subject that has
always given me a headache. I found, as usual, a lot of confusion and
pasting of code. Frustration ensued and I almost moved on. But today I
found a clever piece of code (on TI's chip forum) to calculate this CRC
in software in order to talk to the BSL.

I liked the code so much I created two versions of it - my own
(lightly-edited) C version, and a Forth version. You'll find them here:

  src/crc16.c
  mu/lib/crc16.mu4

In order to try out the C code, put

  LOCALOBJS=	crc16.o

in your src/local.mk (create it if necessary), and run make.

This will create the muforth word  crc16c  which you can test like this:

  token 123456789  crc16c

the result should show on the stack display as 29b1 hex.

To try the Forth version, do this:

  ./muforth -f lib/crc16.mu4

and then

  token 123456789  crc16

Note the lack of a trailing "c" character there... This should yield the
same value.

To see what the FR6989's CRC16 has to say on the subject, load this
awful command line:

  ./muforth -d fr6989 -f target/MSP430/load-kernel.mu4 -f target/MSP430/device/fr6989/crc16.mu4

and then

  chat
  crc-test

You should see the contents of the CRCINIRES (result) and CRCRESR
(reversed result) registers on the stack. The first should be 29b1, as
before.