[core] Renamed and refactored (slightly) the CRC routines
Now the CRC-16 routines (in mu/lib/crc16.mu4 and src/crc16.c) that
process a buffer also take a starting CRC value, instead of always
starting with "ffff.

With this factoring it's possible to mix and match bytes and buffers to
generate a CRC. It doesn't have to be a contiguous chunk of memory.

I didn't change the code that talks to the MSP430 CRC16 hardware,
however. Because of how the peripheral works, it doesn't make as much
sense.