msp430: Support for FR6989's BSL (5xx/6xx protocol) added!
I spent the entire day working on this and there is a huge rant behind
this commit... I'm not going to rant here but it *is* coming...

I'm not even sure how much I want to say here except that you can now
program an FR6989 as long you use something other than the ezFET on the
FR6989 Launchpad! IT DOESN'T FUCKING WORK!!

TI uses weird parity settings in the BSL (for no good reason) and their
own on-board USB debug/UART "solutions" - both on this board and on the
original 430 Launchpad board - can't cope with the BSL's weird parity
issues!! Totally fucking lame and inexcusable.

Anyway, that's a bit of the bigger rant - but it only scratches the
surface of the lameness!!

Argh.