msp430: Huge cleanup of chat, flash, and bsl
The key was changing the interact code and the interface that it
specifies, which each type of debug/flash connection has to implement.

By changing to a simple model that passes around a handful of registers,
reads and writes chunks of memory, and has a simplified interface for
flashing memory, everything got much simpler.

There is a bit more to be done on this score, but this fixes some pretty
warty code that has been bothering me for a long time!