msp430: Fixes to recent chat/flash/bsl commits
I was finally able to test different configurations (chat2/3 X
G2553/FR6989) and found several small bugs, one of which decided
(somehow) to wedge the G2553. In the BSL the only way forward was to
mass erase. I was hoping by using a cleverly-chosen erase command (a504
rather than a506) that the mass erase wouldn't clobber the INFO A memory
(which contains the clock and ADC calibration info!), but as always with
TI, the firmware engineers do their best to screw up as badly as
possible.

With the calibration info gone, it took some time to get the chip
running again. I found some paper notes with values that worked for one
of several G2553 chips. I put some "fail-safe" code into the chat2 and
chat3 firmware that uses those values rather than ffff if the INFO A
page has been erased. It works - with this chip at least. I still need
to do a proper DCO calibration...