ARMv6-M: It's possible to execute Forth words on FRDM-KL25Z !!
By copying - essentially - the memory.mu4, meta.mu4, and interact.mu4
from target/RISC-V, and making a few modifications to the CMSIS-DAP
debug code, the assembler, the disassembler, and by adding a skeleton
Forth kernel (indirect-threaded), it's now possible to execute remote
Forth words on the original Freescale/NXP Freedom board (FRDM-KL25Z).

By doing this is the way that I have - which cleans up the "chat" code
and makes it more modular (and much closer to the style of the MSP430
and RISC-V chat and interact code) I've broken every other ARM platform
the FRDM boards. I need to address how to bring the ST-LINK and TI ICDI
interfaces into line with these changes.