cm0: Fixed two small bugs in asm and disasm with register lists
I caused one of them when creating the macros for the Cortex-M0 (aka
ARMv6-m aka cm0) Forth kernel. I figured you would never put a non-low
register into a reglist, so I made it illegal. But you need to be able
to say

  { lr } push

and

  { pc } pop

Then, in the process of trying to figure out what code was being
compiled for the Forth kernel primitives, I found a bug in the
disassembly of reglists. As mentioned, push and pop can show a register
outside of the r0-r7 range, but ldmia and stmia cannot. But the
disassembler was assuming that bit 8 of the opcode was another register
to display (r0, aka w).

Both now fixed.