riscv: Changed c@ and h@ to *unsigned* and removed *signed* versions
I realized as I was writing the CM0 versions of @ c@ h@ that I had got
this wrong with RISC-V. Signed bytes? They are a pain in the ass! And I
have no idea how people want to the halfword load, but unsigned seems
like a safer bet.

So I switched both c@ and h@ to do zero-extending rather than
sign-extending loads.