[cm3/4] Added support for "bitbanding" in Cortex-M3/M4
"Support" here really means simply a meta-compiler word that will
convert an address (and bit number) in the bitband region into an
address in the alias region.

If you don't know what this means, see the following:

  * Cortex-M3 technical referenc manual, section 4.2, p91 AND
  * target/ARM/v7-m/bitband.mu4

Here is an example of its use. (Connecting to an STM32L-Discovery board,
which has a Cortex-M3 processor.) I've put >>> before what I typed, to
make it clearer.

>>> jtag
ST-LINK/V2  JTAG v14
Debug mode
      SP        RP        IP
200005c8  00000000  00000000*  Ok (chatting) (hex)
  --                 0                 0                 0          decafbad
>>> 2000_1000 du
           v  Q  L  �   L  e     �   �  �  �  �
20001000  76 51 4c dc  4c 65 05 b0  ed ff a5 bd  12 8a 14 04
             dc4c5176     b005654c     bda5ffed     04148a12  Ok (chatting) (hex)
  --                 0                 0          decafbad          20001000
>>> 0 >bitband  Ok (chatting) (hex)
  --                 0                 0          decafbad          22020000
>>> du

22020000  00 00 00 00  01 00 00 00  01 00 00 00  00 00 00 00
             00000000     00000001     00000001     00000000

22020010  01 00 00 00  01 00 00 00  01 00 00 00  00 00 00 00
             00000001     00000001     00000001     00000000

22020020  01 00 00 00  00 00 00 00  00 00 00 00  00 00 00 00
             00000001     00000000     00000000     00000000

22020030  01 00 00 00  00 00 00 00  01 00 00 00  00 00 00 00
             00000001     00000000     00000001     00000000  Ok (chatting) (hex)
  --                 0                 0          decafbad          22020030

Explanation:

First, I picked a random address in the middle of the RAM (2000_1000).
Then I dumped the first 16 bytes/4 words. Then I converted bit 0 of
2000_1000 from a bitband address (because it is in the SRAM bitband
region) into an alias address. Lastly, I dumped the memory at the alias
address.

What you can see from the dumped alias memory is a bit by bit
"unpeeling" of the same data in the first word of the bitband memory.
Note that each *word* in the alias region represents a *bit* in the
bitband region, and that they get peeled off LSB first. So if you read
the words in the alias region from right-to-left they spell out, in
binary, the digits of the first word of the bitband region...

The way I see this being used in muforth is to convert a variable
address or a peripheral register address into an alias address *at
compile time* and then use that constant in an assembler or colon word.

I hope this all makes sense. ;-)