[cm0+] Added support for Freescale's Cortex-M0+ BME
Again, "support" is a big word for something small.

Like the Cortex-M3/4 commit that added "support" for the bitbanding
capability of those processors, this adds support for generating
"decorated load and store" addresses for the Cortex-M0+ processor in
Freescale's (now NXP's) Kinetis L microcontrollers..

Because there is no way, given the ARM v6-M architecture and instruction
set, to do atomic read-modify-write operations, and to save instructions
and time when doing these kinds of operations, Freescale decided to add
a "bit manipulation engine" (BME) to their Cortex-M0+ implementations.

This is similar to, but strictly more powerful than, the bitbanding
support built into the Cortex-M3 and M4 (according to ARM's
documentation, it's optional in the M4). The basic idea is to use
higher-order address bits to annotate (decorate) the semantics of a load
or store operation.

I don't want to repeat here Freescale's documentation or the comments in my
implementation. You should read those yourself:

  * https://www.nxp.com/docs/en/reference-manual/KL25P80M48SF0RM.pdf
    (chapter 17, p271ff)
  * target/ARM/chip/kl25z/bme.mu4