[core] Improved how C code is included for BSDs
Previously, too many BSD systems included usb-freebsd.c, and also the
usb-netbsd.c was always included, because apparently

  #if (__FreeBSD__ < 8)

is *true* if __FreeBSD__ isn't defined! Genius!

The BSD USB code is still incomplete however. This commit doesn't change
that. It can still only talk to USB targets via control requests, which
limits its use with muforth to talking to a JS16 board - of which there
are, to my knowledge, only two, and only one of them is being used with
muforth. ;-)