[hc08] Improved USB chat code - based on new "testbed" experiments
Previously I was doing this crazy thing: storing the negated count of
the bytes to transfer to or from the USB device. It made the code rather
hard to read. I thought it made sense - incrementing a negative 16-bit
count byte-by-byte (using the A register) is easier - at least on the
S08 - than decrementing a positive one. But by switching to using HX and
the AIX instruction, decrementing a positive 16-bit count became easy.
Duh.

By switching to normal, positive counts I was able to save several bytes
of code! And the code is better and easier to read.

Two other changes/additions:

 * the muforth commit that the USB chat code is based on is now compiled
into the flash at fe00, just like with the chat3 serial code. The host
can ask for this chunk of memory at connect time and print it out (which
it now does).

 * the USB vendor ID (VID) and product ID (PID) have changed. I had been
(mis)using Freescale's USB VID (hex 15a2) and just using a "random" PID
(hex c0de) for the JS16 USB chat device. I'm now using a carefully-chosen
VID (f0ff) that suggests how I feel about the USB-IF's Vendor ID program.
Because they are allocating VIDs bottom up (with the exception of Intel,
who of course wanted 8086) it is unlikely that this will conflict any
time soon.