[hc08] JS16 USB chat improved; works with Chromebook's USB3
Previous to this I couldn't use my JS16 board (with homemade USB
firmware) on my Acer Chromebook 14, which has only USB3.0 ports. Oddly
enough, the JS16 worked with everything I tried in on, but failed on the
Chromebook.

I found a race condition in SETUP/OUT token code and fixed it. Now the
board seems to work reliably on the Chromebook! Woot!