[hc08] USB chat can now reliably get a version to host
Before this, the host had to guess blindly that there was a Git commit
stored at fe00. Now it can ask the target for the address of the commit.
It then does a 4-byte read of the memory at that address and prints out
the commit (at "chat" connect time).