[core] Removed src/utf8.c. Use mu/lib/utf8.mu4 instead!
I was frankly surprised by the breakage caused by trying to convert from
UTF-8 to UCS-16 using the tools available in C on *nix/Darwin. Sometimes
a powerful, flexible package (the multibyte/widechar support) is *not*
what you want! Even if you *can* get code that uses it to compile and
run on multiple platforms (which I failed to do).

Rewriting the conversion code in Forth was easy enough, and using it
requires no external libraries or weird environmental dependencies, and
is good enough. In fact, the Forth code does more than the C code! It
also converts from code points to UTF-8!