[core] src/configure.sh smarter about udev rules files
This change might be of interest to Chromebook users - those who have
figured out how to run muforth in a chroot.

Using USB devices with Linux generally means having to futz with udev
rules so that the permissions are correct and you (as a lowly user) can
open the devices you care about.

For a long time the muforth configure script (src/configure.sh) has
created a udev rules file, called 99-muforth.rules. It creates a bunch
of rules that reference the username of the person who ran the script.

In a Chromebook, things are a bit different. The main udevd has no idea
about usernames in your chroot. So we need a rules file that uses UIDs
instead.

I've modified configure.sh to generate this automagically. The UID-based
rules file is called 99-muforth-uid.rules, and the script also spews out
some useful instructions for setting up your rules file(s) in
/run/udev/.

So, head over to muforth/src/ and re-run ./configure.sh