[ARM debug, cm0] Run Forth code with interrupts *off*
Because I've been testing a RAM-resident Forth kernel on a chip with a
program burned into Flash, and all I've done is to *interrupt* said
program, force the chip into debug mode, jam a few values into
registers, and say "now run this code", it's not uncommon for the code
I'm trying to run to get interrupted by the *original* code, jump into a
service routine, and somehow lose its way and never return to debug
mode.

By setting a bit in a debug register (C_MASKINTS in DHCSR) we get better
results.

The *real* solution to this problem is to create our own RAM-resident
vector table, point VTOR at it, set up the processor's mode and both SP
registers, and *then* try executing code. But this stopgap will make it
possible to continue exploring a CMSIS-DAP-based board without too much
pain.