[core] REBUILD: Rename ] to #] and -] to ]
This is a HUGE BREAKING CHANGE.

It's something I've been thinking about for a long time.

-] is a "feature" of muforth that makes it different from other Forths
and confusing to newbies (esp if they have some experience of another
Forth).

Normally, the word  ]  is used to exit interpret mode and enter compile
mode. Because I wanted muforth to "feel" a little bit like colorForth, I
changed the meaning of  ]  to instead do *two* things:

  (1) create a literal from the value on the top of the stack;
  (2) exit interpret mode and return to compile mode.

This means you can do something like this when compiling a word:

  : mi>yd  ( miles -- yards)  [ 5280 3 / ]  * ;

The computation of the number of yards in a mile happens only once, at
compile time, and the  ]  creates a literal with the resulting value.

In most other Forths you would do something like this:

  : mi>yd  ( miles -- yards)  [ 5280 3 / ] literal  * ;

In that case, the  ]  just returns to compile mode; the word  literal
is an immediate (or compiler) word that *executes* at compile time and
creates a literal.

I never liked this, which I why I chose to make  ]  also create a
literal.

This change reverses that decision. Now you have to use  #]  to get the
same behaviour.

literal - now and before - in muforth *does* create a literal, but it is
*not* an immediate/compiler word! So, be careful with that. If you want
the more "standard" behaviour, simply compile this code before loading
code that depends on  literal  being a compiler word:

  compiler
  : literal  literal ;
  forth

Because I was using  ]  as a special word that compiles a literal, I
needed a word that only did the state change. I called that word  -]  .

Since I have now renamed  ]  to  #]  I am free to rename  -]  (my
name for the word that enters compile mode but *doesn't* create a
literal) back to the more normal  ]  .

I did a recursive grep on the code in the muforth tree and hopefully
fixed anything that anyone might be using...